Moved Player into the NavigationBar
Spotlight has been renamed to Pinboard, which contains "Pins" of paper (child components). The Player mockup was transformed into a MiniPlayer which now sits in the NavigationBar. Consequently, the Library selection now propogates up to the Main component and flow down through the NavigationBar to the MiniPlayer for the song info and playback.